chore: shift typings to types
